# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Crusher API
  description: |
    碎石记（Crusher）API 文档

    ## 认证
    大多数端点需要 Bearer Token 认证。Token 通过 Supabase Auth 获取。

    ## 错误响应
    所有错误响应遵循统一格式：
    ```json
    {
      "success": false,
      "error": "错误消息",
      "code": "错误码（可选）",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@crusher.com
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://crusher.com
    description: 生产环境
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: '错误消息'
        code:
          type: string
          example: 'ERROR_CODE'
        details:
          type: object
    ApiMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # 用户相关
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        credits:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # 文档相关
...
    RefundRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        reason:
          type: string
        refund_amount:
          type: integer
        requested_by:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        approved_by:
          type: string
          format: uuid
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # 每日统计相关
    DailyStats:
      type: object
      properties:
        date:
          type: string
          format: date
        new_users:
          type: integer
        orders:
          type: integer
        revenue:
          type: integer
        credits_consumed:
          type: integer
        active_users:
          type: integer
paths:
...
  /api/admin/users/{id}:
    get:
      tags:
        - Admin
      summary: 获取用户详情（管理员）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功返回用户详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Admin
      summary: 更新用户状态（管理员）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - disable_type
              properties:
                disable_type:
                  type: string
                  enum: [normal, login_disabled, usage_disabled]
      responses:
        '200':
          description: 用户状态更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ==================== 管理后台 - 订单 API ====================
...
  /api/admin/refunds:
    get:
      tags:
        - Admin
      summary: 获取退款请求列表（管理员）
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: 成功返回退款请求列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RefundRequest'
                      meta:
                        $ref: '#/components/schemas/ApiMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/refunds/{id}/approve:
    post:
      tags:
        - Admin
      summary: 批准/拒绝退款请求（管理员）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - approved
              properties:
                approved:
                  type: boolean
                rejection_reason:
                  type: string
      responses:
        '200':
          description: 退款请求处理成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RefundRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ==================== 管理后台 - 积分 API ====================
...
  /api/admin/stats:
    get:
      tags:
        - Admin
      summary: 获取今日统计数据（管理员）
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功返回统计数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DailyStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== 认证 API ====================
...
components:
  responses:
    Unauthorized:
      description: 未授权（未登录或 Token 无效）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Unauthorized'
            code: 'UNAUTHORIZED'
    Forbidden:
      description: 禁止访问（权限不足）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Forbidden'
            code: 'FORBIDDEN'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Resource not found'
            code: 'NOT_FOUND'
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Validation failed'
            code: 'VALIDATION_ERROR'
...